refactor(ui): 分离 views 与通用层，业务页面移入 views/ 目录
- sender/, editor/, history/, account_manager/ → views/ 下
- monitor_page.py + monitor_dialog.py + monitor_table.py → views/monitor/
- settings_page.py → views/settings/
- dialogs.py → dialogs/ 包，__init__.py 统一 re-export
- 跨包 import 统一为绝对路径，包内保持相对导入
