+++ src/danmaku_sender/ui/dialogs/__init__.py
@@ -1,0 +1,4 @@
+from .dialogs import AboutDialog, HelpDialog, UpdateDialog, QRLoginDialog
+
+
+__all__ = ["AboutDialog", "HelpDialog", "UpdateDialog", "QRLoginDialog"]
