+++ src/danmaku_sender/ui/views/account/__init__.py
@@ -1,3 +1,4 @@
 from .page import AccountDialog
 
+
 __all__ = ["AccountDialog"]
