+++ src/danmaku_sender/ui/views/editor/components.py
