+++ src/danmaku_sender/ui/views/editor/dialog.py
