+++ src/danmaku_sender/ui/views/monitor/__init__.py
@@ -1,0 +1,5 @@
+from .page import MonitorPage
+from .dialog import MonitorDialog
+
+
+__all__ = ["MonitorPage", "MonitorDialog"]
