+++ src/danmaku_sender/ui/views/monitor/table.py
