+++ src/danmaku_sender/ui/views/sender/__init__.py
@@ -1,3 +1,4 @@
 from .page import SenderPage
 
+
 __all__ = ["SenderPage"]
