+++ src/danmaku_sender/ui/views/sender/components/dialogs/task_builder.py
@@ -1,16 +1,15 @@
 import logging
 
 from PySide6.QtWidgets import (
-    QDialog, QVBoxLayout, QHBoxLayout, QFormLayout, QLabel, QLineEdit,
+    QDialog, QVBoxLayout, QHBoxLayout, QFormLayout, QLineEdit,
     QPushButton, QComboBox, QFileDialog, QMessageBox, QGroupBox
 )
-from PySide6.QtCore import Qt, Signal, Slot
+from PySide6.QtCore import Signal, Slot
 
 from danmaku_sender.controller.video_controller import VideoController
 from danmaku_sender.types.models.video import VideoInfo
 from danmaku_sender.types.models.queue import QueueTask, TaskStatus
 from danmaku_sender.types.models.common import VideoTarget
-from danmaku_sender.config import ApiAuthConfig, SenderConfig
 from danmaku_sender.runtime.state.app_state import AppState
 from danmaku_sender.service.danmaku_parser import DanmakuParser
 from danmaku_sender.utils.string_utils import parse_bilibili_link
