+++ src/danmaku_sender/ui/views/settings/__init__.py
@@ -1,0 +1,4 @@
+from .page import SettingsPage
+
+
+__all__ = ["SettingsPage"]
