+++ src/danmaku_sender/ui/views/settings/page.py
@@ -2,10 +2,8 @@ from PySide6.QtWidgets import (
     QWidget, QVBoxLayout, QHBoxLayout, QFormLayout, QLabel, QCheckBox,
     QGroupBox, QSpinBox, QDoubleSpinBox, QFrame
 )
-from PySide6.QtCore import Qt
 
-from .framework.binder import UIBinder
-
+from danmaku_sender.ui.framework.binder import UIBinder
 from danmaku_sender.runtime.state.app_state import AppState
